Ignore AuditLog in AbpZeroCommonModule
